feat: recipe price 계산
price계산 수정

Resolves #27

+++ backend/src/main/java/com/shinhan/naengtureat/inventory/entity/Inventory.java
@@ -22,7 +22,7 @@ public class Inventory extends SuperEntity {
     private Long id;
 
     @Column(nullable = false)
-    private int quantity;
+    private Double quantity;
 
     @Column(nullable = false, length = 30)
     private String nickName;
